Move scripts to utils
